WIP: Use Data.Vault to store mutable state, not *Ref
The problem with using IORefs and STRefs to store mutable state is
that they are mutable!  To capture the state of a concurrent program,
you need a snapshot of all this state as well as the pure stuff.  But
there was no mechanism to get such a snapshot.

This commit changes the implementation of CRef, MVar, and TVar to use
Data.Vault: a type-safe heterogeneous map.

The vault operations are pure (other than creating a key), so keeping
a snapshot of the single vault value is sufficient to recreate the
entire mutable state of the program.

Despite dragging a lot of state around the computation, this doesn't
have a significant effect on performance.

Profiling results of `dejafu-tests` on my innsmouth:

**before:**
```
total time  =      144.69 secs   (144693 ticks @ 1000 us)
total alloc = 57,918,052,040 bytes
```

**after:**
```
total time  =      147.33 secs   (147333 ticks @ 1000 us)
total alloc = 58,179,981,968 bytes
```

TO DO:

- Maybe tidy up API
- Write changelog entries
- Drop remaining uses of mutable state